+++ openg2p_disbursement_advice/views/openg2p_disbursement_batch_transaction.xml
@@ -2,8 +2,8 @@
 <odoo>
     <data>
         <record id="view_openg2p_disbursement_new_batch" model="ir.ui.view">
-            <field name="name">openg2p.disbursement.new.batch.form</field>
-            <field name="model">openg2p.disbursement.new.batch</field>
+            <field name="name">openg2p.disbursement.batch.transaction.form</field>
+            <field name="model">openg2p.disbursement.batch.transaction</field>
             <field name="arch" type="xml">
                 <form string="Batch Form">
                     <header>
@@ -17,7 +17,6 @@
                             <label for="name" class="oe_edit_only"/>
                             <h1>
                                 <field name="name"/>
-                                <!-- <field name="state" invisible="1"/> -->
                             </h1>
                          </div>
                         <group>
@@ -43,13 +42,15 @@
         </record>
     </data>
     <record id="view_openg2p_disbursement_main_tree" model="ir.ui.view">
-        <field name="name">openg2p.disbursement.new.batch.tree</field>
-        <field name="model">openg2p.disbursement.new.batch</field>
+        <field name="name">openg2p.disbursement.batch.transaction.tree</field>
+        <field name="view_mode">tree</field>
+        <field name="model">openg2p.disbursement.batch.transaction</field>
         <field name="arch" type="xml">
             <tree>
                 <field name="name"/>
                 <field name="program_id"/>
-                <field name="state" invisible="1"/>
+                <field name="state"/>
+                <button name="action_confirm" string="Confirm" class="oe_highlight"/>
             </tree>
         </field>
     </record>
