+++ openg2p_disbursement_advice/views/openg2p_disbursement_main.xml
@@ -16,12 +16,7 @@
                             </group>
                             <group>
                                 <field name="state" invisible="1"/>
-                                <field name="acc_holder_name"/>
-                                <field name="bank_account_id"/>
-                                <field name="name"/>
                                 <field name="amount"/>
-                                <field name="payment_mode"/>
-                                <field name="currency_id"/>
                             </group>
                             <group name="secondary">
                                 <label for="date_start" string="Period"/>
@@ -47,9 +42,8 @@
         <field name="arch" type="xml">
             <tree>
                 <field name="beneficiary_id"/>
-                <field name="acc_holder_name"/>
-                <field name="name"/>
                 <field name="batch_id"/>
+                <field name="amount"/>
             </tree>
         </field>
     </record>
