+++ openg2p_disbursement_advice/views/res_partner_bank.xml
@@ -14,6 +14,7 @@
                                 <field name="sequence" invisible="1"/>
                                 <field name="bank_id" options="{'no_create_edit': True, 'no_quick_create': True}"/>
                                 <field name="acc_number"/>
+                                <field name="beneficiary_id"/>
                                 <field name="payment_mode"/>
                                 <field name="company_id" groups="base.group_multi_company"
                                        options="{'no_create_edit': True, 'no_quick_create': True}"/>
